classification_model/requirements.txt
cookiecutter==1.7.3
ipykernel==6.9.1
ipython==8.0.1
feature-engine==1.2.0
joblib==1.1.0
matplotlib==3.5.1
numpy==1.22.2
pandas==1.4.1
scikit-learn==1.0.2
scipy==1.8.0
statsmodels==0.13.2
PyYAML==5.4.1